Added a function that tests whether a bucketlist item exists before adding it

+++ myapp/models/bucketlist_item.py
@@ -20,10 +20,22 @@ class BucketListItem(BaseModel):
     completed = db.Column(db.Boolean, default=False)
 
 
+    def __item_exists(self, item, bucketlist_id):
+        """
+        Method is used to verify that a bucketlist name exists in
+        the database
+        """
+        if BucketListItem.query.filter_by(bucketlist_id=bucketlist_id).filter(BucketListItem.name.ilike(item)).first():
+            return True
+        return False
+
+
     def save_bucketlist_item(self):
         """
         Method is used to add a bucketlist item to the database
         """
+        if self.__item_exists(self.name, self.bucketlist_id):
+            return False
         self.add_data_set()
         return True
 
@@ -32,7 +44,8 @@ class BucketListItem(BaseModel):
         """
         Method is used to add a bucketlist item to the database
         """
+        if not self.__item_exists(self.name, self.bucketlist_id):
+            return False
         self.delete_data_set()
-        db.session.commit()
         return True
